added check for column width Std Dev, changed findColumnBreaks to return boolean
added the Std Dev check to Find Column Breaks method. Also changed the
return type to be boolean to say if Find Column breaks was successful.
Segment Image helper function now takes in boolean from find columns
and either continues the segmentation if true is found or if false it
will output the image name to a fail images text file.