fix: FullPage is passed, but no snippet is generated.
The issue was raised when the snippen's width is longer than fullpage's. For example, where no column break is found, becuase of poor image quality and severe skew, the entire page is considered as a snippet. Then, the snippet's height is calculated as following: 14/9 * snippet_width, which is longer than the fullpage's.

In order to solve this problem, we check if snippet's height is less than or equal to the fullpage's height. Otherwise, the image will not be passed, and error will be logged in imageFailedNeedHuman.txt file.